+++ src/bin/investigate.ts
@@ -1,10 +1,10 @@
 import 'source-map-support/register';
+import { engineConfigFromENV } from '../lib/config/env';
 import DataDir from '../lib/data/dir';
 import { DataSet } from '../lib/data/set';
 import { Engine } from "../lib/engine/engine";
 import { printSummary } from "../lib/engine/summary";
 import { Hubspot } from '../lib/hubspot';
-import { engineConfigFromENV } from '../lib/parameters/env-config';
 
 const engine = new Engine(Hubspot.memory(), engineConfigFromENV());
 const data = new DataSet(DataDir.root.subdir('in')).load();
