+++ src/bin/run-once.ts
@@ -1,10 +1,10 @@
 import 'source-map-support/register';
+import { engineConfigFromENV } from '../lib/config/env';
 import DataDir from '../lib/data/dir';
 import { DataSet } from '../lib/data/set';
 import { Engine } from "../lib/engine/engine";
 import { Hubspot } from '../lib/hubspot';
 import { Logger } from '../lib/log';
-import { engineConfigFromENV } from '../lib/parameters/env-config';
 
 const dataDir = DataDir.root.subdir('in');
 const log = new Logger(dataDir.subdir(`once-${Date.now()}`));
