+++ src/bin/save-attribution-csv.ts
@@ -1,9 +1,9 @@
 import 'source-map-support/register';
+import { engineConfigFromENV } from '../lib/config/env';
 import DataDir from '../lib/data/dir';
 import { DataSet } from '../lib/data/set';
 import { Engine } from "../lib/engine/engine";
 import { Hubspot } from '../lib/hubspot';
-import { engineConfigFromENV } from '../lib/parameters/env-config';
 import { isPresent, sorter } from "../lib/util/helpers";
 
 const engine = new Engine(Hubspot.memory(), engineConfigFromENV());
