+++ src/lib/hubspot/index.ts
@@ -1,5 +1,5 @@
+import { hubspotContactConfigFromENV, hubspotDealConfigFromENV } from "../config/env";
 import { Logger } from "../log";
-import { hubspotContactConfigFromENV, hubspotDealConfigFromENV } from "../parameters/env-config";
 import { CompanyManager } from "./model/company";
 import { ContactManager, HubspotContactConfig } from "./model/contact";
 import { DealManager, HubspotDealConfig } from "./model/deal";
