+++ src/lib/marketplace/api.ts
@@ -1,7 +1,7 @@
 import got from 'got';
 import { DateTime, Duration, Interval } from 'luxon';
+import { mpacCredsFromENV } from '../config/env';
 import { Progress } from '../log/download';
-import { mpacCredsFromENV } from '../parameters/env-config';
 import { AttachableError, KnownError } from '../util/errors';
 import { RawLicense, RawTransaction } from './raw';
